refactoring: move formula parsing outside Syft::ExplicitStateDfaMona::dfa_of_formula
The proposed refactoring aims to make parsing separated from the Syft::ExplicitStateDfaMona::dfa_of_formula function.

The function now only deals with translation from logic to DFA. The parsing of the formula is delegated to the client code.

This is a preparatory change for applying pre-processing techniques for solving the synthesis task.