feat: optimize one-step unreal-check for F(a) and F(~a)
In case the argument of the eventually operator is a literal, we can
improve the unrealizability check by checking whether the variable of the literal
is controlled or not. Therefore, instead of returning "true", we return "a" (or "~a"),
such that the checker can take such literal into account.

This is very useful for patterns of the type:

F(a) & F(b) & F(c) & ...

where the variables are all uncontrollable. The previous version of the check
would have considered the formula as "don't know", whereas the new extended check
would consider it as unrealizable.